front: move fork source lookup into conversation resource
